changes for python3 but will run with python2
