+++ dla_cnn/data_model/Sightline.py
@@ -1,0 +1,79 @@
+import numpy as np
+
+
+class Sightline(object):
+
+    def __init__(self, id, dlas=None, flux=None, loglam=None, z_qso=None):
+        self.flux = flux
+        self.loglam = loglam
+        self.id = id
+        self.dlas = dlas
+        self.z_qso = z_qso
+        self.data_markers = []     # An array of DataMarker objects which mark areas of the sightline
+        self.prediction = None
+
+
+    # Returns the data in the legacy data1, qso_z format for code that hasn't been updated to the new format yet
+    def get_legacy_data1_format(self):
+        raw_data = {}
+        raw_data['flux'] = self.flux
+        raw_data['loglam'] = self.loglam
+        raw_data['plate'] = self.id.plate if hasattr(self.id, 'plate') else 0
+        raw_data['mjd'] = self.id.mjd if hasattr(self.id, 'mjd') else 0
+        raw_data['fiber'] = self.id.fiber if hasattr(self.id, 'fiber') else 0
+        raw_data['ra'] = self.id.ra if hasattr(self.id, 'ra') else 0
+        raw_data['dec'] = self.id.dec if hasattr(self.id, 'dec') else 0
+        return raw_data, self.z_qso
+
+
+    # Clears all fields of the DLA
+    def clear(self):
+        self.flux = None
+        self.loglam = None
+        self.id = None
+        self.dlas = None
+        self.z_qso = None
+        self.prediction = None
+        self.data_markers = []
+
+
+    def is_lyb(self, peakix):
+        """
+        Returns true if the given peakix (from peaks_ixs) is the ly-b of another DLA in the set peaks_ixs in prediction
+        :param peakix:
+        :return: boolean
+        """
+        assert self.prediction is not None and peakix in self.prediction.peaks_ixs
+
+        lambda_higher = (10**self.loglam[peakix]) / (1025.722/1215.67)
+
+        # An array of how close each peak is to beign the ly-b of peakix in spectrum reference frame
+        peak_difference_spectrum = np.abs(10**self.loglam[self.prediction.peaks_ixs] - lambda_higher)
+        nearest_peak_ix = np.argmin(peak_difference_spectrum)
+
+        # get the column density of the identfied nearest peak
+        _, potential_lya_nhi, _, _ = \
+            self.prediction.get_coldensity_for_peak(self.prediction.peaks_ixs[nearest_peak_ix])
+        _, potential_lyb_nhi, _, _ = \
+            self.prediction.get_coldensity_for_peak(peakix)
+
+        # Validations: check that the nearest peak is close enough to match
+        #              sanity check that the LyB is at least 0.3 less than the DLA
+        is_nearest_peak_within_range = peak_difference_spectrum[nearest_peak_ix] <= 15
+        is_nearest_peak_larger_coldensity = potential_lyb_nhi < potential_lya_nhi - 0.3
+
+        return is_nearest_peak_within_range and is_nearest_peak_larger_coldensity
+
+
+    def get_lyb_index(self, peakix):
+        """
+        Returns the index location of the Ly-B absorption for a given peak index value
+        :param peakix:
+        :return: index location of Ly-B
+        """
+        spectrum_higher = 10**self.loglam[peakix]
+        spectrum_lambda_lower = spectrum_higher * (1025.722 / 1215.67)
+        log_lambda_lower = np.log10(spectrum_lambda_lower)
+        ix_lambda_lower = (np.abs(self.loglam - log_lambda_lower)).argmin()
+        return ix_lambda_lower
+
