+++ dla_cnn/gen_catalogs.sh
