+++ dla_cnn/preprocess.sh
