+++ dla_cnn/training_set.py
