--- src/data_model/Dla.py
@@ -1,5 +1,0 @@
-
-class Dla():
-    def __init__(self, central_wavelength, col_density=0):
-        self.central_wavelength = central_wavelength        # spectrum
-        self.col_density = col_density
